+++ simulator/UT0311L04/UTC0311L04.go
@@ -189,6 +189,9 @@ func (s *UT0311L04) Handle(src *net.UDPAddr, rq messages.Request) {
 	case *messages.ActivateAccessKeypadsRequest:
 		s.activateKeypads(src, v)
 
+	case *messages.RestoreDefaultParametersRequest:
+		s.restoreDefaultParameters(src, v)
+
 	default:
 		panic(fmt.Errorf("unsupported message type %T", v))
 	}